service view and edit temp fix
